+++ tests/PuliPluginTest.php
@@ -434,7 +434,7 @@ class PuliPluginTest extends JsonWriterTestCase
         $this->assertJsonFileEquals($this->tempDir.'/puli-all-installed.json', $this->tempDir.'/puli.json');
     }
 
-    public function testWarnIfPackageNotLoadable()
+    public function testWarnIfPackageNotInstallable()
     {
         $event = new CommandEvent(ScriptEvents::POST_INSTALL_CMD, $this->composer, $this->io);
 
@@ -462,6 +462,32 @@ class PuliPluginTest extends JsonWriterTestCase
         $this->assertFileExists($this->tempDir.'/repository');
 
         $this->assertJsonFileEquals($this->tempDir.'/puli-partially-installed.json', $this->tempDir.'/puli.json');
+    }
+
+    public function testWarnIfPackageNotLoadable()
+    {
+        $event = new CommandEvent(ScriptEvents::POST_INSTALL_CMD, $this->composer, $this->io);
+
+        copy($this->tempDir.'/puli-not-loadable.json', $this->tempDir.'/puli.json');
+
+        $this->installPaths['vendor/package1'] = $this->tempDir.'/not-loadable';
+
+        $this->io->expects($this->at(0))
+            ->method('write')
+            ->with('<info>Looking for updated Puli packages</info>');
+        $this->io->expects($this->at(1))
+            ->method('write')
+            ->with('<warning>Warning: Could not load vendor/package1 (not-loadable): UnsupportedVersionException: Cannot read package file not-loadable/puli.json at version 5.0. The highest readable version is 1.0. Please upgrade Puli.</warning>');
+        $this->io->expects($this->at(2))
+            ->method('write')
+            ->with('<info>Building Puli resource repository</info>');
+
+        $this->plugin->postInstall($event);
+
+        $this->assertFileExists($this->tempDir.'/My/PuliFactory.php');
+        $this->assertFileExists($this->tempDir.'/repository');
+
+        $this->assertJsonFileEquals($this->tempDir.'/puli-not-loadable.json', $this->tempDir.'/puli.json');
     }
 
     public function testCopyComposerPackageNameToPuli()