+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-* 1.0.0-next (@release_date@)
+* 1.0.0-beta2 (2015-01-27)
 
  * fixed: packages with a moved install path are reinstalled now
  * added `IOLogger`